split out sidebar
